:recycle: refactor: moving const variables to data.ts
