--- src/lib/components/dialog/dialog.utils.ts
@@ -1,6 +1,0 @@
-export const COLOR_MAP: Record<string, string> = {
-  warning: 'warning',
-  info: 'primary',
-  success: 'success',
-  error: 'error',
-};
